chore: upgrade to Vue3 (refs #1325 #315 #1226 #1520)
Change build and serve process. It runs in vite
Add upgrade event in Editor.vue and Viewer.vue for handle content changes
