--- apps/vue-editor/demo/esm/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head lang="en">
-    <meta charset="UTF-8" />
-    <title>Demo</title>
-  </head>
-  <body>
-    <div id="editor"></div>
-    <!-- Editor -->
-    <script type="module" src="./index.js"></script>
-  </body>
-</html>
