--- apps/vue-editor/snowpack.config.js
@@ -1,11 +1,0 @@
-/** @type {import("snowpack").SnowpackUserConfig } */
-module.exports = {
-  mount: {
-    'demo/esm': '/',
-    src: '/dist',
-  },
-  devOptions: {
-    port: 8080,
-  },
-  plugins: ['@morgul/snowpack-plugin-vue2'],
-};
