+++ apps/vue-editor/src/main.js
@@ -1,0 +1,6 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+
+const app = createApp(App);
+
+app.mount('#app')
