+++ apps/vue-editor/webpack.config.js
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const path = require('path');
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const {VueLoaderPlugin} = require('vue-loader');
 const webpack = require('webpack');
 const { version, author, license } = require('./package.json');
